Added autodetection of SBT path, improved console, other fixes
SBT path is now autodetected on first startup and in case the path has
been set to an empty string.

The double-tab null pointer exception has been fixed.

The ShellConsole has been moved to a separate package.

The console should now move the caret to the end after executing a
command more often, however this is still not fully resolved yet.
Support for improvements has been added - the console is now notified
whenever the OutputStream is modified, ie. when the SBT process writes
data.

SBTUtils improvements related to SBT information.